Split prefill and decode batchers to separate implementations
Prefill and decode should be batched separately as their preferred batch
sizes for each piece of hardware.
